+++ changelog.md
@@ -1,4 +1,4 @@
-13.1.0
+13.2.0
   - added clientId to query websocket url
   - added equity curve filter to CopyFactory
   - fixed health state tracking for multiple replicas